Code formatting and cleanup - format source code - rename the filename parser to CCWarcFilenameParser   (it parses WARC file names/paths but not the index)
